geting errors right

+++ server/src/main/java/service/AuthService.java
@@ -1,6 +1,7 @@
 package service;
 
 import dataaccess.AuthDataAccess;
+import dataaccess.DBException;
 import dataaccess.DataAccessException;
 import model.AuthData;
 
@@ -19,23 +20,34 @@ public class AuthService {
         }
     }
 
-    public AuthData getAuth(String authToken) throws UserException, AuthException {
+    public AuthData getAuth(String authToken) throws DataAccessException {
         try {
             if(authToken == null){
                 return null;
             }
             AuthData auth = authDB.getAuth(authToken);
             return auth;
-        }catch (DataAccessException e){
-            throw new AuthException("Error: Unauthorized");
+        }catch (DBException e){
+            throw new DataAccessException(e.getMessage());
+            //throw new AuthException("Error: Unauthorized");
+        } catch (DataAccessException e) {
+            throw new DataAccessException("Error: " + e.getMessage());
         }
     }
 
     public void deleteAuth(String authToken) throws DataAccessException {
         try {
+            AuthData auth = getAuth(authToken);
+            if(auth == null){
+                throw new AuthException("Error: You are not Logged in");
+            }
             authDB.deleteAuth(authToken);
+        }catch (AuthException e){
+            throw new AuthException(e.getMessage());
         }catch (DataAccessException e){
-            throw new AuthException("Error: You are not Logged in");
+            throw new DataAccessException(e.getMessage());
+        } catch (DBException e) {
+            throw new DataAccessException(e.getMessage());
         }
     }
